feat: create Home component and import it into index.tsx
