+++ src/components/Home.tsx
@@ -1,0 +1,19 @@
+import Head from "next/head";
+import Layout from "./Layout";
+
+export default function Home() {
+  return (
+    <>
+      <Head>
+        <title>SIML.AI</title>
+        <meta name="description" content="Landing page for SIML.AI" />
+        <meta name="viewport" content="width=device-width, initial-scale=1" />
+        <link rel="icon" href="/favicon.ico" />
+      </Head>
+
+      <Layout>
+        
+      </Layout>
+    </>
+  );
+}
